+++ bin/index.js
@@ -7,7 +7,6 @@ const program = require('commander');
 const pkg = require('../package.json');
 
 const OpenAPI = require(path.resolve(__dirname, '../dist/index.js'));
-const { rmdirSync } = require('fs');
 
 const params = program
     .name('openapi')
@@ -15,6 +14,10 @@ const params = program
     .version(pkg.version)
     .option('-i, --input <value>', 'OpenAPI specification, can be a path, url or string content (required)')
     .option('-o, --output <value>', 'Output directory (required)')
+    .option('-oc, --outputCore <value>', 'Output directory for core files')
+    .option('-os, --outputServices <value>', 'Output directory for services')
+    .option('-om, --outputModels <value>', 'Output directory for models')
+    .option('-osm, --outputSchemas <value>', 'Output directory for schemas')
     .option('-c, --client <value>', 'HTTP client to generate [fetch, xhr, node]', 'fetch')
     .option('--useOptions <value>', 'Use options instead of arguments', false)
     .option('--useUnionTypes <value>', 'Use union types instead of enums', false)
@@ -24,6 +27,9 @@ const params = program
     .option('--exportSchemas <value>', 'Write schemas to disk', false)
     .option('--clean <value>', 'Clean a directory before generation', true)
     .option('--request <value>', 'Path to custom request file')
+    .option('--interfacePrefix <value>', 'Prefix for interface model(default: "I")')
+    .option('--enumPrefix <value>', 'Prefix for enum model(default: "E")')
+    .option('--typePrefix <value>', 'Prefix for type model(default: "T")')
     .parse(process.argv)
     .opts();
 
@@ -36,24 +42,69 @@ function isValidJson(value) {
     }
 }
 
-function generate(config) {
-    return OpenAPI.generate({
-        input: config.input,
-        output: config.output,
-        httpClient: config.client,
-        useOptions: isValidJson(config.useOptions) ? JSON.parse(config.useOptions) : false,
-        useUnionTypes: isValidJson(config.useUnionTypes) ? JSON.parse(config.useUnionTypes) : false,
-        exportCore: isValidJson(config.exportCore) ? JSON.parse(config.exportCore) : false,
-        exportServices: isValidJson(config.exportServices) ? JSON.parse(config.exportServices) : false,
-        exportModels: isValidJson(config.exportModels) ? JSON.parse(config.exportModels) : false,
-        exportSchemas: isValidJson(config.exportSchemas) ? JSON.parse(config.exportSchemas) : false,
-        clean: isValidJson(config.clean) ? JSON.parse(config.clean) : false,
-        request: config.request,
-    })
+/**
+ * Preparation the configuration from the root level
+ */
+function prepareRootOptions(options) {
+    return {
+        output: options.output,
+        outputCore: options.outputCore,
+        outputServices: options.outputServices,
+        outputModels: options.outputModels,
+        outputSchemas: options.outputSchemas,
+        httpClient: options.client,
+        useOptions: isValidJson(options.useOptions) ? JSON.parse(options.useOptions) : false,
+        useUnionTypes: isValidJson(options.useUnionTypes) ? JSON.parse(options.useUnionTypes) : false,
+        exportCore: isValidJson(options.exportCore) ? JSON.parse(options.exportCore) : false,
+        exportServices: isValidJson(options.exportServices) ? JSON.parse(options.exportServices) : false,
+        exportModels: isValidJson(options.exportModels) ? JSON.parse(options.exportModels) : false,
+        exportSchemas: isValidJson(options.exportSchemas) ? JSON.parse(options.exportSchemas) : false,
+        clean: isValidJson(options.clean) ? JSON.parse(options.clean) : false,
+        request: options.request,
+        interfacePrefix: options.interfacePrefix ? options.interfacePrefix : 'I',
+        enumPrefix: options.enumPrefix ? options.enumPrefix : 'E',
+        typePrefix: options.typePrefix ? options.typePrefix : 'T',
+    };
+}
+
+/**
+ * Preparation the configuration from second level
+ */
+function prepareOptions(options, rootOptions) {
+    return options.map(option => {
+        return {
+            input: option.input,
+            output: option.output ? option.output : rootOptions ? rootOptions.output : '',
+            outputCore: option.outputCore ? option.outputCore : rootOptions ? rootOptions.outputCore : '',
+            outputServices: option.outputServices ? option.outputServices : rootOptions ? rootOptions.outputServices : '',
+            outputModels: option.outputModels ? option.outputModels : rootOptions ? rootOptions.outputModels : '',
+            outputSchemas: option.outputSchemas ? option.outputSchemas : rootOptions ? rootOptions.outputSchemas : '',
+            httpClient: option.client ? option.client : rootOptions ? rootOptions.client : '',
+            useOptions: isValidJson(option.useOptions) ? JSON.parse(option.useOptions) : rootOptions ? rootOptions.useOptions : false,
+            useUnionTypes: isValidJson(option.useUnionTypes) ? JSON.parse(option.useUnionTypes) : rootOptions ? rootOptions.useUnionTypes : false,
+            exportCore: isValidJson(option.exportCore) ? JSON.parse(option.exportCore) : rootOptions ? rootOptions.exportCore : false,
+            exportServices: isValidJson(option.exportServices) ? JSON.parse(option.exportServices) : rootOptions ? rootOptions.exportServices : false,
+            exportModels: isValidJson(option.exportModels) ? JSON.parse(option.exportModels) : rootOptions ? rootOptions.exportModels : false,
+            exportSchemas: isValidJson(option.exportSchemas) ? JSON.parse(option.exportSchemas) : rootOptions ? rootOptions.exportSchemas : false,
+            clean: isValidJson(option.clean) ? JSON.parse(option.clean) : rootOptions ? rootOptions.clean : false,
+            request: option.request ? option.request : rootOptions ? rootOptions.request : '',
+            interfacePrefix: option.interfacePrefix ? option.interfacePrefix : rootOptions ? rootOptions.interfacePrefix : 'I',
+            enumPrefix: option.enumPrefix ? option.enumPrefix : rootOptions ? rootOptions.enumPrefix : 'E',
+            typePrefix: option.typePrefix ? option.typePrefix : rootOptions ? rootOptions.typePrefix : 'T',
+        };
+    });
+}
+
+function generate(options) {
+    return OpenAPI.generate(options)
         .then(() => {
-            console.group(`Generation from "${config.input}"`);
-            console.log(`Generation from "${config.input}" was finished\n`);
-            console.log(`Output folder: ${path.resolve(process.cwd(), config.output)}\n`);
+            console.group(`Generation from has been finished`);
+            const group = Array.isArray(options) ? options : Array.of(options);
+            group.forEach(option => {
+                console.log(`Generation from "${option.input}" was finished`);
+                console.log(`Output folder: ${path.resolve(process.cwd(), option.output)}`);
+                console.log('==================================');
+            });
             console.groupEnd();
         })
         .catch(error => {
@@ -68,15 +119,9 @@ if (OpenAPI) {
     if (fs.existsSync(configFile)) {
         const dataString = fs.readFileSync(configFile, { encoding: `UTF-8` });
         const configs = isValidJson(dataString) ? JSON.parse(dataString) : [];
-        if (params.clean) {
-            configs.forEach(config => {
-                rmdirSync(config.output, { recursive: true, force: true });
-            });
-        }
-        configs.forEach(config => {
-            generate(config);
-        });
+        let operations = Array.isArray(configs) ? Array.of(prepareOptions(configs)) : prepareOptions(configs.items, prepareRootOptions(configs));
+        generate(operations);
     } else {
-        generate(params);
+        generate(prepareOptions(params));
     }
 }