+++ src/openApi/v2/parser/getOperationParameters.ts
@@ -2,7 +2,6 @@ import type { OperationParameters } from '../../../client/interfaces/OperationPa
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiParameter } from '../interfaces/OpenApiParameter';
 import { Parser } from '../Parser';
-import { getOperationParameter } from './getOperationParameter';
 import { sortByRequired } from './sortByRequired';
 
 export function getOperationParameters(this: Parser, openApi: OpenApi, parameters: OpenApiParameter[]): OperationParameters {
@@ -20,7 +19,7 @@ export function getOperationParameters(this: Parser, openApi: OpenApi, parameter
     // Iterate over the parameters
     parameters.forEach(parameterOrReference => {
         const parameterDef = (parameterOrReference.$ref ? this.context.get(parameterOrReference.$ref) : parameterOrReference) as OpenApiParameter;
-        const parameter = getOperationParameter(openApi, parameterDef);
+        const parameter = this.getOperationParameter(openApi, parameterDef);
 
         // We ignore the "api-version" param, since we do not want to add this
         // as the first / default parameter for each of the service calls.