+++ src/openApi/v3/parser/getType.spec.ts
@@ -1,12 +1,13 @@
-import { getType, GetTypeName } from './getType';
-import { Parser } from '../Parser';
-import { Context } from '../../../core/Context';
 import RefParser from 'json-schema-ref-parser';
+
+import { Context } from '../../../core/Context';
+import { Parser } from '../Parser';
+import { getType } from './getType';
 
 describe('getType', () => {
     it('should convert int', async () => {
         const parser = new RefParser();
-        const context = new Context('test/spec/v3.yml');
+        const context = new Context('test/spec/v3.yml', { output: './generated' });
         context.addRefs(await parser.resolve('test/spec/v3.yml'));
         const type = new Parser(context).getType('int', '');
         expect(type.type).toEqual('number');
@@ -17,7 +18,7 @@ describe('getType', () => {
 
     it('should support file with ext', async () => {
         const parser = new RefParser();
-        const context = new Context('test/spec/v3.yml');
+        const context = new Context('test/spec/v3.yml', { output: './generated' });
         context.addRefs(await parser.resolve('test/spec/v3.yml'));
         const type = new Parser(context).getType('schemas/ModelWithString.yml', '');
         expect(type.type).toEqual('IModelWithString');
@@ -44,7 +45,7 @@ describe('getType', () => {
 
     it('should support external generation type', async () => {
         const parser = new RefParser();
-        const context = new Context(object);
+        const context = new Context(object, { output: './generated' });
         context.addRefs(await parser.resolve(object));
         const type = new Parser(context).getType('#/components/schemas/someSpecialSchema', '');
         expect(type.type).toEqual('ISomeSpecialSchema');