+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.00-beta] - 2021-12-11
+
+### Added
+
+- Add setting up the tag to the publication(CI)
+- The common configuration is moved from the array to the root block
+- Output paths for models, services, schemas and core can be set up separately
+
+### Fixed
+
+- Bug with an incorrect prefix which set to an enum if one has property - "type"
+
 ## [0.1.32] - 2021-12-03
 
 ### Fixed